requirements.txt
requests
nose
requirements-parser